+++ tests/test_router.py
@@ -3,9 +3,9 @@ import json
 
 import pytest
 
-from bitbucket_webhooks_router import event_schemas
-from bitbucket_webhooks_router import hooks
-from bitbucket_webhooks_router import router
+from bitbucket_webhooks import event_schemas
+from bitbucket_webhooks import hooks
+from bitbucket_webhooks import router
 
 
 def test_no_handler_available() -> None:
